make in same phase with previous step
