--- src/services/actions/actions.js
@@ -1,19 +1,0 @@
-export const addTodo = (props) => {
-  const { text } = props;
-  return {
-    type: "ADD",
-    payload: {
-      text: text,
-    },
-  };
-};
-
-
-export const deleteTodo = (id) => {
-  return {
-    type: "DELETE",
-    payload: {
-      id: id,
-    },
-  };
-};
