+++ src/services/actions/todo.js
@@ -1,6 +1,15 @@
 export const addTodo = (props) => {
-  return {
-    type: "ADD",
-    payload: props,
-  };
+    return {
+        type: "ADD",
+        payload: props,
+    };
+};
+
+export const deleteTodo = (id) => {
+    return {
+        type: "DELETE",
+        payload: {
+            id: id,
+        },
+    };
 };
