+++ src/services/reducers/todo.js
@@ -1,20 +1,26 @@
 const initialTodos = {
-  todos: [],
+    todos: [],
 };
 
 const todoReducer = (state = initialTodos, action) => {
-  switch (action.type) {
-    case "ADD":
-      return {
-        ...state,
-        todos: [
-          action.payload,
-          ...state.todos,
-        ],
-      };
-    default:
-      return state;
-  }
+    switch (action.type) {
+        case "ADD":
+            return {
+                ...state,
+                todos: [
+                    action.payload,
+                    ...state.todos,
+                ],
+            };
+        case "DELETE":
+            const updatedTodos = state.todos.filter((todo) => todo.id !== action.payload.id);
+            return {
+                ...state,
+                todos: updatedTodos,
+            };
+        default:
+            return state;
+    }
 };
 
 export default todoReducer;
